requirements.txt
adagio==0.2.4
aiohttp==3.8.4
aiosignal==1.3.1
ansi2html==1.8.0
antlr4-python3-runtime==4.11.1
appdirs==1.4.4
arviz
async-timeout==4.0.2
asynctest==0.13.0
attrs==23.1.0
backcall==0.2.0
Bottleneck
catboost==1.2
certifi 
cffi 
cftime 
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
cmdstanpy==1.1.0
colorama 
convertdate==2.4.0
cycler 
Cython==0.29.35
darts==0.24.0
dash==2.10.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.7
decorator==5.1.1
entrypoints==0.4
ephem==4.1.4
Flask==2.2.5
fonttools==4.38.0
frozenlist==1.3.3
fs==2.4.16
fsspec==2023.1.0
fugue==0.8.5
fugue-sql-antlr==0.1.6
future 
gevent==22.10.2
graphviz==0.20.1
greenlet==2.0.2
holidays==0.27.1
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
ipykernel==6.16.2
ipython==7.34.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter-dash==0.4.2
jupyter_client==7.4.9
jupyter_core==4.12.0
kiwisolver
lazy_loader==0.2
lightgbm==3.3.5
lightning-utilities==0.8.0
llvmlite==0.39.1
LunarCalendar==0.0.9
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib 
matplotlib-inline==0.1.6
mkl-service==2.4.0
multidict==6.0.4
multipledispatch==0.6.0
munkres==1.1.4
natsort==8.4.0
nest-asyncio==1.5.6
netCDF4 
nfoursid==1.0.1
numba==0.56.4
numexpr 
numpy
opt-einsum 
orbit-ml 
orjson==3.9.1
packaging 
pandas 
pandas-datareader==0.10.0
pandas-flavor==0.5.0
parso==0.8.3
patsy 
pickleshare==0.7.5
Pillow==9.5.0
plotly==5.15.0
plotly-resampler==0.8.3.2
pmdarima==2.0.3
prompt-toolkit==3.0.38
prophet==1.1.4
protobuf==4.23.3
psutil==5.9.5
pyarrow==12.0.1
pycparser
Pygments==2.15.1
pyjanitor==0.23.1
PyMeeus==0.5.12
pyod==1.0.9
pyodbc==4.0.39
pyparsing
pyro-api 
pyro-ppl 
pystan==2.19.1.1
python-dateutil 
pytictoc==1.5.2
pytorch-lightning==1.9.5
pytz 
pywin32==306
PyYAML
pyzmq==25.1.0
qpd==0.4.3
requests==2.31.0
retrying==1.3.4
scikit-learn==1.0.2
scipy 
seaborn
shap==0.41.0
six 
slicer==0.0.7
SQLAlchemy==1.4.45
sqlglot==16.4.0
statsforecast==1.5.0
statsmodels
tbats==1.1.3
tenacity==8.2.2
tensorboardX==2.6.1
threadpoolctl==3.1.0
torch==1.13.1
torchmetrics==0.11.4
tornado
tqdm 
trace-updater==0.0.9.1
traitlets==5.9.0
triad==0.9.0
typing_extensions==4.6.3
urllib3==2.0.3
wcwidth==0.2.6
Werkzeug==2.2.3
wincertstore==0.2
wsgi-oauth2==0.2.2
xarray==0.20.2
xgboost==1.6.2
yarl==1.9.2
zipp==3.15.0
zope.event==4.6
zope.interface==6.0